Separate columns for SPI
Previously the str and hex columns showed SPI data from the MOSI and
MISO line in an interleaved fashion. This is rarely useful.

Also, due to a bug, for SPI the hex column started with the string
"unknown error" even though no actual error occurred.

This commit separates MOSI and MISO data into separate columns and also
fixes the bug.